Jobber på forum.. (forms)
git-svn-id: https://subversion.kantega.no/svn/aksess/projects/aksess-modules@1895 497bca9e-fae7-0310-9681-f2327195b1a6

+++ modules/forum/src/java/no/kantega/forum/model/Post.java
@@ -13,7 +13,7 @@ import java.util.Set;
 public class Post {
     private long Id;
     private ForumThread Thread;
-    private int ReplyToId;
+    private long ReplyToId;
     private String Subject;
     private String Body;
     private User Owner;
@@ -28,11 +28,11 @@ public class Post {
         this.Id = id;
     }
 
-    public int getReplyToId() {
+    public long getReplyToId() {
         return ReplyToId;
     }
 
-    public void setReplyToId(int replyToId) {
+    public void setReplyToId(long replyToId) {
         ReplyToId = replyToId;
     }
